requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
facebook-sdk==3.1.0
instagrapi==2.0.0
python-dotenv==1.0.0
schedule==1.2.0
jinja2==3.1.2
markdown==3.5.1
pillow==10.1.0
selenium==4.15.2
webdriver-manager==4.0.1
sqlalchemy==2.0.23
aiohttp==3.9.1
python-dateutil==2.8.2
pytz==2023.3
tweepy==4.14.0
flask==3.0.0
flask-limiter==3.5.0
flask-wtf==1.2.1
bcrypt==4.1.2
pytest==7.4.3